pyproject.toml
[project]
name = "mcp-doctor-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "google-api-python-client>=2.179.0",
    "google-auth>=2.49.2",
    "mcp>=1.27.0",
    "openai>=2.31.0",
    "pydantic>=2.13.0",
    "python-dotenv>=1.2.2",
    "resend>=2.28.0",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.44.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.10",
]
